tools/data_types/validation_types.py: Renamed from validation/data.py; no functional change
